moved the handleHomeClock funcxtion into the context page
